requirements.txt
jupyter
pandas==1.2.4
ipython==7.22.0
apache-airflow[aws,amazon,postgres]==2.2.3
boto3~=1.18.65
numpy==1.19.5
apache-airflow-providers-amazon==2.6.0
botocore==1.21.65
Unidecode==1.3.2
geopandas==0.10.2
paramiko===2.4.2